# Disable prepared statements via pgx config