Fixes https://github.com/authzed/spicedb/issues/1217

Instead of addressing this issue by documenting configuration required for pgbouncer to work, this PR removes the need for special pgbouncer configuration.

In the past, [a bug was found](https://spicedb.dev/d/force-custom-plan) with the use of prepared statements which impacted SpiceDB's performance. SpiceDB uses pgx to interact with Postgres which [uses prepared statements by default](https://github.com/jackc/pgx/wiki/Automatic-Prepared-Statement-Caching).

SpiceDB addressed the previously mentioned performance issue by injecting the `plan_cache_mode=force_custom_plan` setting into the postgres connection string. This fix mimicked the fix put in place by the author of [the blog post](https://spicedb.dev/d/force-custom-plan) outlining the original investigation into the bug.

While this fix was the only immediate option for the blog post author, SpiceDB can do better by changing pgx's execution mode. The use of `plan_cache_mode` does not truly disable prepared statements, but instead leaves pgx continuing its use of prepared statements while SpiceDB simultaneously asks postgres to ignore the query plans it computes for them upfront. Disabling prepared statements entirely via pgx has several benefits:
* If SpiceDB decides to selectively enable prepared statements for some queries in the future then it will be possible.
* Not disabling prepared statements via `plan_cache_mode` simplifies the story around PgBouncer deployments by removing the need  for special PgBouncer configuration. There are currently two PgBouncer configurations that are needed to use PgBouncer with SpiceDB:
  - `POOL_MODE=session` - This is needed to deal with [an issue reported by a user](https://github.com/authzed/spicedb/issues/1217) wherein PgBouncer was cleaning up prepared statements. (An odd issue to encounter given SpiceDB wishes to disable prepared statements.)
  - `IGNORE_STARTUP_PARAMETERS=plan_cache_mode` - PgBouncer does not understand the `plan_cache_mode` parameter, so it must be instructed to ignore it and pass it through to Postgres. This is another strange requirement. SpiceDB's method of disabling prepared statements is leaking out to PgBouncer users.

Discord conversation: https://discord.com/channels/844600078504951838/900405749405089812/1164379648650649700