@bradengroom 2023-10-31
Still WIP. I plan to write some tests for `defaultQueryExecMode`. For now, I think it's closer to what Victor suggested.