@bradengroom 2023-11-02
@vroldanbet I ran the benchmarks on both this branch and main to make sure I'm using up to date results.

this branch:
```
BenchmarkDatastoreDriver/postgres/TestTuple/SnapshotRead-10                                     	    2560	    521779 ns/op	   12890 B/op	     224 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadOnlyNamespace/BySubject-10        	      58	  20676485 ns/op	 2376906 B/op	   55159 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadOnlyNamespace/ByResource-10       	     100	  12058841 ns/op	 2376806 B/op	   55159 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadWithRelation/ByResource-10        	     338	   3385922 ns/op	 2378896 B/op	   55180 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadWithRelation/BySubject-10         	      78	  20775982 ns/op	 2378724 B/op	   55179 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadAllResourceFields/ByResource-10   	    2240	    521134 ns/op	   16757 B/op	     286 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadAllResourceFields/BySubject-10    	    2314	    538459 ns/op	   16758 B/op	     286 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SnapshotReverseRead-10                              	     100	  13424350 ns/op	 2377927 B/op	   55174 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/Touch-10                                            	     861	   1492052 ns/op	   12263 B/op	     185 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/Create-10                                           	     562	   1788423 ns/op	   10136 B/op	     147 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/0_-10                                	      24	  71060014 ns/op	12762846 B/op	  145667 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/0.1_-10                              	      15	  69570836 ns/op	12262022 B/op	  136301 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/0.25_-10                             	      16	  70275073 ns/op	10857274 B/op	  122497 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/0.5_-10                              	      20	  61499912 ns/op	 8762044 B/op	   99247 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/1_-10                                	      37	  40974766 ns/op	 5452780 B/op	   52962 allocs/op
```

main:
```
BenchmarkDatastoreDriver/postgres/TestTuple/SnapshotRead-10                                     	    2408	    462334 ns/op	   12890 B/op	     224 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadOnlyNamespace/BySubject-10        	      63	  20798763 ns/op	 2376729 B/op	   55159 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadOnlyNamespace/ByResource-10       	      84	  11989514 ns/op	 2376877 B/op	   55159 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadWithRelation/ByResource-10        	     356	   3298096 ns/op	 2378879 B/op	   55180 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadWithRelation/BySubject-10         	      79	  20965164 ns/op	 2378815 B/op	   55180 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadAllResourceFields/ByResource-10   	    2192	    500163 ns/op	   16758 B/op	     286 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SortedSnapshotReadAllResourceFields/BySubject-10    	    2118	    490966 ns/op	   16756 B/op	     286 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/SnapshotReverseRead-10                              	      96	  13861234 ns/op	 2377958 B/op	   55173 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/Touch-10                                            	     829	   1412580 ns/op	   12402 B/op	     185 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/Create-10                                           	     889	   1402953 ns/op	   10200 B/op	     144 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/0_-10                                	      24	  48112885 ns/op	12167486 B/op	  145690 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/0.1_-10                              	      28	  43083798 ns/op	11611001 B/op	  136315 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/0.25_-10                             	      25	  43459102 ns/op	10314110 B/op	  122515 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/0.5_-10                              	      26	  46815375 ns/op	 8213800 B/op	   99264 allocs/op
BenchmarkDatastoreDriver/postgres/TestTuple/CreateAndTouch/1_-10                                	      37	  34119407 ns/op	 5110080 B/op	   52965 allocs/op
```

The largest difference is on the `CreateAndTouch` tests. `CreateAndTouch/0_-10` is registering as 22ms slower. I'm unsure if this is an acceptable regression or not (I assume not). It's worth noting that removing the use prepared statements will definitely be slower (by _some_ amount). Even though SpiceDB currently disables the use of cached plans for prepared statements, it is likely still benefiting from prepared statements in these ways:
* Fewer bytes over the wire (pgx only has to send `EXECUTE foo(args1, arg2, ...);` instead of the fully formed statement). This also incurs less of a parsing cost on the postgres side since there's just less to parse.
* Likely no statement -> logical plan conversion. Even though we've asked postgres not to reuse the same _physical_ plan, I would not be surprised if it is still reusing the same _logical_ plan.

While the benefits of turning off prepared statements provides a streamlined deployment story for pgbouncer users:
* No restriction to use the session pool_mode
* No need to ignore the plan_cache_mode setting

So I think that's the trade-off in this PR: perf vs ease of pgbouncer deployment. Given the nature of the project, I assume perf wins. If the Authzed folks agree then I'm happy to close this PR and go back to documenting the required PgBouncer configs. This was an interesting exercise regardless 🙂

I think it's not _as_ strange to document the pgbouncer config requirements now that they can be tied to perf improvements.