@bradengroom 2023-11-02
It's also worth noting that the `CreateAndTouch` test creates 1000 tuples:
https://github.com/authzed/spicedb/blob/018aad9be9172e9c57999bdf3cd52c112e0bb00f/internal/datastore/benchmark/driver_bench_test.go#L199

So that 22 additional milliseconds is spread out over 1k operations. (22 _microseconds_ per operation). After realizing that, it doesn't actually feel that bad I think I would go for this. I'm still to pursue whatever the Authzed folks want though 🙂