@bradengroom 2023-11-02
Sorry, one last comment. If this gets merged, it can be disabled by adding the following to your postgres connection string:
`?default_query_exec_mode=cache_statement&plan_cache_mode=force_custom_plan`