@vroldanbet 2023-11-02
Hey @bradengroom, thank you so much for the thorough investigation, this is excellent work 💯 . I'm a bit on the fence here - I recall also looking into disabling prepared statements back in the day and I recall that it regressed somewhat the latencies. I'm going to run some tests internally to see what it looks like. Let's hold on merging this for now 🙏🏻